Add ssim2qff mock utility (#1)
We have added a first mock utility (ssim2qff) using a simple approach based in string matching. There is also a shell script that runs the mock utility tests and can also be used as a usage example. The shell script is located in the 'tests' folder. Currently the mock utility doesn't support log messages, but in the future it is possible to add more exit codes in order to track down different types of errors.