Add SHA256 checksum to verify the input files (#3)
We check that the SHA256 checksum of each of the given input files matches the checksum of the expected ones. We add the same functionality to the example utilities flights and ssim2qff. This ensures the fast check of the input files, as the real ones may be too large to verify with a file comparison approach. We are using a base 16 representation of the checksums and storing them as strings, because we think this makes the code more readable.