Add tests for the mock executable (#10)
Add a shell script that runs the mock executable with different arguments to check that behaves as expected. The Makefile rule test-mockexec runs the shell script and outputs an error message if it does not pass some test.
Change the temporary files to the tmp directory because it is more organised and easier to clean. The executables are still created inside their package folder.