Add temporary directory (#11)
Add tmp directory.
The .gitignore file is used to be able to add an empty directory to the repository.
Right now, the tests create temporary test data files and need the directory to exist. The data files are not created inside test/testdata because they are really simple and temporary.