Add function to generate configuration files (#2)
Add function that generates a Protocol Buffer configuration file from command line arguments.
This function allows to easily create a configuration file from the expected command line arguments. The type of each argument needs to be added before the argument. Instead of getting a name of a configuration file as input and writing the data to the file in that path, we print the result directly to the standard output to gain more flexibility to the user.
We also add tests for this function.